# Revamp support for long windows