--- a/ovgenpy/channel.py
+++ b/ovgenpy/channel.py
@@ -1,6 +1,7 @@
 from typing import TYPE_CHECKING, Any
 
-from ovgenpy.config import register_config
+from ovgenpy.config import register_config, Alias
+from ovgenpy.util import coalesce
 from ovgenpy.wave import _WaveConfig, Wave
 
 
@@ -15,11 +16,16 @@ class ChannelConfig:
 
     trigger: 'ITriggerConfig' = None    # TODO test channel-specific triggers
     # Multiplies how wide the window is, in milliseconds.
-    trigger_width: int = 1
-    render_width: int = 1
+    trigger_width: int = None
+    render_width: int = None
 
     ampl_ratio: float = 1.0     # TODO use amplification = None instead?
     line_color: Any = None
+
+    # region Legacy Fields
+    trigger_width_ratio = Alias('trigger_width')
+    render_width_ratio = Alias('render_width')
+    # endregion
 
 
 class Channel:
@@ -39,8 +45,11 @@ class Channel:
         self.wave = Wave(wcfg, cfg.wav_path)
 
         # Compute subsampling (array stride).
-        self.trigger_subsampling = subsampling * cfg.trigger_width
-        self.render_subsampling = subsampling * cfg.render_width
+        tw = coalesce(cfg.trigger_width, ovgen_cfg.trigger_width)
+        self.trigger_subsampling = subsampling * tw
+
+        rw = coalesce(cfg.render_width, ovgen_cfg.render_width)
+        self.render_subsampling = subsampling * rw
 
         # Compute window_samp and tsamp_frame.
         nsamp = ovgen_cfg.render_width_s * self.wave.smp_s / subsampling