… .idea/inspectionProfiles/Project_Default.xml
… ovgenpy/utils/windows.py
… tests/test_channel.py
… tests/test_config.py
… tests/test_trigger.py
… tests/test_windows.py