- (Mostly) fix lag when using long input windows
    - Prevent CorrelationTrigger from looking backwards >= 1 frame in time)
    - (Don't want to implement) configurable left-taper steepness, up to half frame?
    - Maybe removing 2-frame repetition is a mistake when dealing with <60Hz waves
- Add `trigger_diameter` field (* window size = movement distance)
- Add `trigger_falloff[0,1]` fields (* audio period = correlation diameter and taper)
- Rename fields and convert old names into aliases
- Move trigger defaults from `default_config()` to `CorrelationTriggerConfig`

Fixes #17.

Potential issues:
- Does it hurt treble quality?
- Performance impact?
    - Don't calculate pitch twice.
- [x] Reexamine 01469c1de1340a7293eed679ea1d8d5370cabbeb
    - it's not a Gaussian anymore, it's a `min(half-Hann with configurable right end, pitch-dependent envelope)`.
- [x] Drastically hurts quality of Time Trax bell bass (https://www.youtube.com/watch?v=DcwqdeJ7DSk)
    - PR fixes lag but reduces usable data width. Fixed with extra trigger_width.
- [x] Performance: Benchmarking and profiling
------
sd3-winter trigger benchmark
- Before: 388fps
- After: 196fps

- [x] Calculate pitch once, pass into _update_buffer
    - 289fps
- [x] Don't regenerate window unless period changes by at least 1 semitone
    - ~~361fps~~ BUG: no data windowing whatsoever
    - 332fps